reorg eql
