+++ specification/eql/search/EqlSearchRequest.ts
@@ -22,6 +22,7 @@ import { ExpandWildcards, Field, IndexName } from '@_types/common'
 import { float, uint } from '@_types/Numeric'
 import { QueryContainer } from '@_types/query_dsl/abstractions/container/QueryContainer'
 import { Time } from '@_types/Time'
+import { EqlResultPosition, EqlSearchFieldFormatted } from './types'
 
 /**
  * @rest_spec_name eql.search
@@ -108,28 +109,4 @@ export interface Request extends RequestBase {
      */
     result_position?: EqlResultPosition
   }
-}
-export class EqlSearchFieldFormatted {
-  /**
-   * Wildcard pattern. The request returns values for field names matching this pattern.
-   */
-  field: Field
-  /**
-   * Format in which the values are returned.
-   */
-  format?: string
-}
-
-/**
- * Set of matching events or sequences to return.
- */
-export enum EqlResultPosition {
-  /**
-   * Return the most recent matches, similar to the Unix tail command.
-   */
-  tail = 0,
-  /**
-   * Return the earliest matches, similar to the Unix head command.
-   */
-  head = 1
 }
